clean and some fix for multi
