+++ Django_data/static/js/Game/Game_Events.js
@@ -48,11 +48,12 @@ function SetTheGame (event) {
   console.log(data.message);
   gameCanvas.inGame = true;
   if (data.message === "Game stopped") {
+	EndGame("Game already finished!");
 	if (deleteEvent === true) {
 	  return;
 	}
 	gameSocket.removeEventListener("message", SetTheGame);
-	EndGame("You won!\nPlayer has leaved the game");
+	// EndGame("You won!\nPlayer has leaved the game");
   } else {
 	if (deleteEvent === true) {
 	  return;
