+++ Services/PongServer/ServerFile/global_var.py
@@ -1,0 +1,5 @@
+CONNECTIONS_GAMES = dict()
+# ws : game_id (int)
+
+GAMES = dict()
+# game_id (int) : {"players" : [(username,websocket), (username, websocket)], "settings" : dict, "nb_players" : int, "pong" : Pong, player_1 : user, player_2 : user}
